+++ Package.swift
@@ -3,8 +3,10 @@ import PackageDescription
 
 #if os(macOS)
 private let addCryptoSwift = false
+private let binaryPlugin = true
 #else
 private let addCryptoSwift = true
+private let binaryPlugin = false
 #endif
 
 let frameworkDependencies: [Target.Dependency] = [
@@ -39,7 +41,7 @@ let package = Package(
             name: "SwiftLintPlugin",
             capability: .buildTool(),
             dependencies: [
-                .target(name: "swiftlint")
+                .target(name: binaryPlugin ? "SwiftLintBinary" : "swiftlint")
             ]
         ),
         .executableTarget(
@@ -98,5 +100,10 @@ let package = Package(
                 "SwiftLintTestHelpers"
             ]
         ),
+        .binaryTarget(
+            name: "SwiftLintBinary",
+            url: "https://github.com/realm/SwiftLint/releases/download/0.50.1/SwiftLintBinary-macos.artifactbundle.zip",
+            checksum: "487c57b5a39b80d64a20a2d052312c3f5ff1a4ea28e3cf5556e43c5b9a184c0c"
+        )
     ]
 )